Void command metasource edge case fix
if an entry has a command is intended as a no-input void method, while the entry also has a non-null metasource or meta input or both, previously we would fail gracefully by not running the method, which isn't exactly helping, so instead we fail gracefully by not failing gracefully and runing the method anyway.

in any other case where a method expects inputs and the metasource or metainput are null, it will fail gracefully by not running the method and pushing a warning. fair?